third_party/WebKit/LayoutTests/fast/html/navigation-transition.html

<!DOCTYPE html>

<meta name="transition-elements" content="#transitionElement,#boxTransitionElement;*">
<script src="../../resources/js-test.js"></script>
<script>

description('Tests for the HTML markup serializer used for transition elements and transition element names and rects');
jsTestIsAsync = true;

window.onload = function() {
    if (window.internals) {
        var serializedHTML = internals.serializeNavigationMarkup();
        var testiframe = document.getElementById('testiframe');
        testiframe.contentDocument.write(serializedHTML);

        transitionElement = document.getElementById('transitionElement');
        sourceElement = document.getElementById('boxTransitionElement');
        destElement = testiframe.contentDocument.getElementById('boxTransitionElement');

        sourceElementRect = sourceElement.getBoundingClientRect();

        bottomElement = document.getElementById('bottomElement');

        preHidingBottomElementRect = bottomElement.getBoundingClientRect();
        preHidingTransitionElementOpacity = window.getComputedStyle(transitionElement).opacity;

        internals.hideAllTransitionElements();

        postHidingTransitionElementOpacity = window.getComputedStyle(transitionElement).opacity;
        postHidingBottomElementRect = bottomElement.getBoundingClientRect();

        shouldBe('sourceElementRect.top', 'destElement.offsetTop');
        shouldBe('sourceElementRect.left', 'destElement.offsetLeft');
        shouldBe('sourceElementRect.height', 'destElement.offsetHeight');
        shouldBe('sourceElementRect.width', 'destElement.offsetWidth');

        shouldBeTrue('preHidingTransitionElementOpacity == 1');
        shouldBeTrue('postHidingTransitionElementOpacity == 0');

        // Hiding transition elements should not shift any other elements on the page.
        shouldBe('preHidingBottomElementRect.top', 'postHidingBottomElementRect.top');
        shouldBe('preHidingBottomElementRect.left', 'postHidingBottomElementRect.left');

        transitionElementIds = internals.getTransitionElementIds();
        shouldBeEqualToString('transitionElementIds[0]', 'transitionElement');
        shouldBeEqualToString('transitionElementIds[1]', 'boxTransitionElement');

        transitionElementRects = internals.getTransitionElementRects();
        transitionElementBounds = internals.boundsInRootViewSpace(transitionElement);
        boxTransitionElementBounds = internals.boundsInRootViewSpace(sourceElement);
        shouldBeTrue('transitionElementRects[0].left == transitionElementBounds.left');
        shouldBeTrue('transitionElementRects[0].top == transitionElementBounds.top');
        shouldBeTrue('transitionElementRects[0].width == transitionElementBounds.width');
        shouldBeTrue('transitionElementRects[0].height == transitionElementBounds.height');
        shouldBeTrue('transitionElementRects[1].left == boxTransitionElementBounds.left');
        shouldBeTrue('transitionElementRects[1].top == boxTransitionElementBounds.top');
        shouldBeTrue('transitionElementRects[1].width == boxTransitionElementBounds.width');
        shouldBeTrue('transitionElementRects[1].height == boxTransitionElementBounds.height');

        internals.showAllTransitionElements();
        postShowingTransitionElementOpacity = window.getComputedStyle(transitionElement).opacity;
        shouldBeTrue('postShowingTransitionElementOpacity == 1');

        finishJSTest();
    }
}
</script>

<style>

body {
    background: white;
    margin-height: 0px;
    margin-width: 0px;
}

#wrappingElement {
    background: gray;
    padding: 13px 20px 8px 30px;
    border-style: solid;
    border-width: 5px;
    width: 400px;
    margin: 44px 2px 13px 30px;
}

#transitionElement {
    padding-bottom: 0px;
}

#boxTransitionElement {
    background: red;
    padding: 8px 50px 50px 18px;
    border-style: solid;
    border-width: 30px;
    margin: 55px 15px 25px 135px;
}

#blueBox {
    padding: 25px 75px 50px 100px;
}

#testiframe {
    width: 100%;
    height: 600px;
    background: gray;
    border-width: 0px;
}

</style>

<div id="wrappingElement">
    <div id="transitionElement">Header</div>

    <div id="boxTransitionElement">
        <img id="blueBox" src="resources/images/blue.png">
    </div>

    <div id="bottomElement">Footer</div>
</div>

<iframe id="testiframe"></iframe>
