third_party/WebKit/LayoutTests/fast/html/navigation-transition-expected.txt

Tests for the HTML markup serializer used for transition elements and transition element names and rects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sourceElementRect.top is destElement.offsetTop
PASS sourceElementRect.left is destElement.offsetLeft
PASS sourceElementRect.height is destElement.offsetHeight
PASS sourceElementRect.width is destElement.offsetWidth
PASS preHidingTransitionElementOpacity == 1 is true
PASS postHidingTransitionElementOpacity == 0 is true
PASS preHidingBottomElementRect.top is postHidingBottomElementRect.top
PASS preHidingBottomElementRect.left is postHidingBottomElementRect.left
PASS transitionElementIds[0] is "transitionElement"
PASS transitionElementIds[1] is "boxTransitionElement"
PASS transitionElementRects[0].left == transitionElementBounds.left is true
PASS transitionElementRects[0].top == transitionElementBounds.top is true
PASS transitionElementRects[0].width == transitionElementBounds.width is true
PASS transitionElementRects[0].height == transitionElementBounds.height is true
PASS transitionElementRects[1].left == boxTransitionElementBounds.left is true
PASS transitionElementRects[1].top == boxTransitionElementBounds.top is true
PASS transitionElementRects[1].width == boxTransitionElementBounds.width is true
PASS transitionElementRects[1].height == boxTransitionElementBounds.height is true
PASS postShowingTransitionElementOpacity == 1 is true
PASS successfullyParsed is true

TEST COMPLETE
Header

Footer
